+++ packages/renderer/src/lib/chat/components/chat.svelte
@@ -117,10 +117,12 @@ let attachments = $state<Attachment[]>([]);
 			{readonly}
 			loading={chatClient.status === 'streaming' || chatClient.status === 'submitted'}
 			messages={chatClient.messages}
+      {selectedModel}
+      {selectedMCP}
 		/>
      <form class="bg-background mx-auto flex w-full gap-2 px-4 pb-4 md:max-w-3xl md:pb-6">
        {#if !readonly}
-         <MultimodalInput {attachments} {user} {chatClient} selectedModel={selectedModel} selectedMCP={selectedMCP} class="flex-1" />
+         <MultimodalInput {attachments} {user} {chatClient} class="flex-1" />
        {/if}
      </form>
    </div>
